==> digital.sexing.cells/datasets/2019_bulk_data/05_SVM_on_embryo.R <==

library(ggplot2)
library(gridExtra)
#################################################
## read in meta information
df=read.table('GSE127176_series_matrix.txt',skip=30,sep='\t',fill=TRUE)
head(df)
df=t(df[c(1,10,39),])
df=df[-1,]
colnames(df)=c('sample.id','stage','GSM.id')
df.meta=df;
dim(df.meta) #54 samples
###################################################
## read in STAR output tab-limited read count data
files=Sys.glob('GSE127176_RAW/*tab')
files
raw.expr=list();
for(file in files){
  gsm=strsplit(basename(file),'_')[[1]][1]
  x=read.table(file,skip=4)
  raw.expr[[gsm]]=x
}
length(raw.expr) #54

# extract 'expressed genes', defined by >=21 counts in >=5 embryos
per.embryo.gene=lapply(raw.expr,function(x){
  x[x[,2]>=21,1]
})
sapply(per.embryo.gene,length)
expr.genes=names(which(table(unlist(per.embryo.gene))>=5))
length(expr.genes) #8983 genes
filtered.expr=lapply(raw.expr,function(x){
  x[x[,1] %in% expr.genes,]
})
sapply(filtered.expr,nrow) #8983 genes
#################################################
## check for chr Y genes
df.gene.table=data.table::fread('validate.id_2019_paper_data.txt')
head(df.gene.table)
table(df.gene.table$chromosome_name)
# no gene on chrY 

#################################################
## TPM: read.count/gene.length
## then divided/ sum(length.corrented.count) x 10^6
df.info=data.table::fread('dmel_geneLength_chr.txt')
#query.genes=scan('all.gene.names.txt',what='')
query.genes=expr.genes
head(query.genes)
length(query.genes) #8983
sum(df.info$GENEID %in% query.genes) #8934
tmp=merge(filtered.expr[[1]],df.info,by.x='V1',by.y='GENEID')
length.corrected.counts=lapply(filtered.expr,function(x){
  tmp=merge(x,df.info,by.x='V1',by.y='GENEID')
  tmp$V2/tmp$exonic.gene.sizes/1000 #in kilobase
})
sapply(length.corrected.counts,length)
TPM=lapply(length.corrected.counts,function(x){
  x/sum(x)*10^6
})
log2.TPM=lapply(TPM,function(i) log(i+1,base=2))
#################################################
## PCA on log2(TPM)
log2TPM.df=Reduce(`cbind`,log2.TPM)
dim(log2TPM.df) #8934 x 54
rownames(log2TPM.df)=tmp[,1]
colnames(log2TPM.df)=names(log2.TPM)
log2TPM.df[1:3,1:3]

# remove genes with var larger than 90% quantile
all.genes.var=apply(log2TPM.df,1,var)
quantile(all.genes.var,0.90)
log2TPM.df.genes=log2TPM.df[all.genes.var<quantile(all.genes.var,0.90),]
dim(log2TPM.df.genes) #8040   54

x=colnames(log2TPM.df.genes)
rownames(df.meta)=df.meta[,3]
df.meta=as.data.frame(df.meta[x,])
sum(df.meta[,3]==colnames(log2TPM.df.genes)) #54
df.meta.keep=df.meta;

library(factoextra)
res.pca=prcomp(t(log2TPM.df.genes),scale=T) #PCA using all filtered genes 8040 

head((res.pca$sdev^2)/sum(res.pca$sdev^2))
var.explained=res.pca$sdev^2/sum(res.pca$sdev^2)
var.explained=round(var.explained*100,2)

df.meta$stage=as.numeric(gsub('Stage: Stage ','',df.meta$stage))
basic_plot <- fviz_pca_ind(res.pca, label="none")
ggplot(cbind(basic_plot$data,df.meta),
       aes(x=x,y=y,col=factor(stage)))+geom_point(shape=19)+theme_bw()+
  xlab(paste0('PC1 (',var.explained[1],'%)'))+
  ylab(paste0('PC2 (',var.explained[2],'%)'))

##############################################
## sex samples using k-means clustering
df.gene.table=data.table::fread('validate.id_2019_paper_data.txt')
pick.genes1=df.gene.table[grep('sxl|msl|mof|mle|roX',ignore.case = T,df.gene.table$SYMBOL),]
pick.genes1; #8 genes
#"mle" "msl-3"  "msl-2"   "msl-1" "mof" "lncRNA:roX2" "lncRNA:roX1" 'Sxl'
#markers=pick.genes1$FLYBASE

# only use Sxl, msl-2, roX1, rox2
#markers=c('FBgn0264270','FBgn0005616','FBgn0019661','FBgn0019660')
markers=c('FBgn0264270','FBgn0005616','FBgn0019661')
sum( markers %in% expr.genes) #4
sub.df=log2TPM.df[rownames(log2TPM.df) %in% markers,]
dim(sub.df) #4 x 54 samples

if(!file.exists('sample.meta_sex.label.txt')){
  df=scale(t(sub.df)) #sample by feature matrix
  head(df)
  out=kmeans(df, centers=3, iter.max = 10, nstart = 1)
  table(out$cluster)
  # 1  2  3 
  # 20 29  5
  
  sum(df.meta$GSM.id==names(out$cluster))
  df.meta$cluster=out$cluster
  table(df.meta$stage,df.meta$cluster)
  # compare with embr201948138-sup-0006-tableev1.xlsx
  x=factor(df.meta$cluster);
  levels(x)=c('female','PB','male')
  df.meta$cluster=x
  table(df.meta$cluster)
  head((res.pca$sdev^2)/sum(res.pca$sdev^2))
  
  pdf('all.stages.pdf',useDingbats = T,height=4)
  basic_plot <- fviz_pca_ind(res.pca, label="none")
  ggplot(cbind(basic_plot$data,df.meta),
         aes(x=x,y=y,col=factor(stage),shape=cluster))+
    geom_point(size=3)+theme_bw()+
    xlab(paste0('PC1 (',var.explained[1],'%)'))+
    ylab(paste0('PC2 (',var.explained[2],'%)'))
  dev.off()
  
  data.table::fwrite(df.meta,'sample.meta_sex.label.txt')
}
df.meta=data.table::fread('sample.meta_sex.label.txt')
table(df.meta$cluster)

## svm 
library(e1071)

#sub.df=log2TPM.df
rownames(sub.df)
dim(sub.df) #4 gene, 54 sample

sum(colnames(sub.df)==df.meta$GSM.id)
head(df.meta)

sub.df=sub.df[,df.meta$cluster!='PB']
df.meta=df.meta[df.meta$cluster!='PB',]
sum(colnames(sub.df)==df.meta$GSM.id) #49 samples
df.meta$cluster.id=as.numeric(factor(df.meta$cluster))
table(df.meta$cluster.id) #1=female, 2=male

data = data.frame(t(sub.df), y =df.meta$cluster.id)
head(data)

data.svm = svm(y ~ ., data = data, kernel = "linear", cost = 10, scale = FALSE)
print(data.svm)

plot(data.svm, data)

library(caTools)
dataset=data
dim(dataset) #49 sample x 5 feature gene
#dataset=data[,-3] #remove roX2
colnames(dataset)
set.seed(123)
#set.seed(321) 
split = sample.split(dataset$y, SplitRatio = 0.7)

training_set = subset(dataset, split == TRUE)
test_set = subset(dataset, split == FALSE)
(label.col=ncol(dataset)) #5nd column is the cluster label

# Feature Scaling for train and test separately
training_set[-label.col] = scale(training_set[-label.col])
test_set[-label.col] = scale(test_set[-label.col])

classifier = svm(formula = y ~ .,
                 data = training_set,
                 type = 'C-classification',
                 kernel = 'linear')

# Predicting the Test set results
y_pred = predict(classifier, newdata = test_set[-label.col])
table(test_set$y,y_pred) 


# use all 2019 data to construct classifier Predicting the Test set results
if(T){
  library(org.Dm.eg.db)
  dim(dataset) #49 x 5
  x=AnnotationDbi::select(org.Dm.eg.db,keys=colnames(dataset)[-label.col],
                          keytype='FLYBASE',columns=c('SYMBOL'))
  x$SYMBOL #"msl-2"       "lncRNA:roX1" "Sxl"  
  colnames(dataset)[1:(label.col-1)]=x$SYMBOL
  dim(dataset)
  scaled.dataset= scale(dataset[-label.col])
  scaled.dataset=cbind(scaled.dataset,dataset$y)
  colnames(scaled.dataset)[label.col]='y'
  #scaled.dataset$y=as.data.frame(scaled.dataset$y)
  classifier = svm(formula = y ~ .,
                   data = scaled.dataset,
                   type = 'C-classification',
                   kernel = 'linear')
  y_pred = predict(classifier, newdata = scaled.dataset)
  table(scaled.dataset[,label.col],y_pred)
}
saveRDS(classifier,'svm_classifier_3genes_all49samples.rds')


## extend to single cell data
## read in embryo data
x=readRDS('../2019_paper_reproduce.result/prepared_embryo.data.rds')
test=as.matrix(x$umi.mat)
test.meta=x$sample.meta
class(test)

gene.names=rownames(test);
pick.genes=gene.names[grep('Sxl|msl|roX|mle|mof',gene.names)]
pick.genes=gene.names[grep('Sxl|msl_2|roX',gene.names)]
test=test[pick.genes,]
dim(test)
test_set=t(test)
test_set = scale(test_set)
colnames(test_set)
colnames(test_set)[1]='msl-2'
#colnames(test_set)=c('msl.2','roX1','roX2','Sxl')
#test_set=test_set[,-3] #remove roX2


y_pred = predict(classifier, newdata = test_set)
table(y_pred)
#y_pred
#1   2 
#816 481  
table(y_pred,test.meta$sex)
